Move to (app) grouped folder
