+++ src/app/(app)/dashboard/layout.js
@@ -1,0 +1,9 @@
+export default function Layout({ children, header }) {
+    return (
+        <>
+            {header}
+
+            <main>{children}</main>
+        </>
+    )
+}
