+++ src/app/(app)/layout.js
@@ -1,18 +1,16 @@
 'use client'
 
-import Navigation from '@/components/Layouts/Navigation'
 import { useAuth } from '@/hooks/auth'
+import Navigation from '@/app/(app)/Navigation'
 
-const AppLayout = ({ children, header }) => {
+const AppLayout = ({ children }) => {
     const { user } = useAuth({ middleware: 'auth' })
 
     return (
         <div className="min-h-screen bg-gray-100">
             <Navigation user={user} />
 
-            {header}
-
-            <main>{children}</main>
+            {children}
         </div>
     )
 }
